rename util package
